Actualizar carga de datos en dataset.py

+++ dataset.py
@@ -177,8 +177,11 @@ class MRIDataset(Dataset):
         super().__init__()
         self.data_dir = data_dir
         self.split = split
-        self.data = torch.from_numpy(np.load(f"{self.data_dir}{self.split}.npy")) 
-        self.labels = torch.from_numpy(np.load(f"{self.data_dir}{self.split}_labels.npy"))
+       
+
+        self.data = torch.from_numpy(np.load(f"{self.data_dir}{self.split}.npy")).float()
+        self.labels = torch.from_numpy(np.load(f"{self.data_dir}{self.split}_labels.npy")).float()
+
